requirements.txt
selenium==4.16.0
Pillow==10.1.0
pytesseract==0.3.10
webdriver_manager==4.0.1
pandas==2.2.0
